Completed Checkout Button, refactored Floats, and remodeled seed

+++ server/db/models/product.js
@@ -9,7 +9,7 @@ const Product = db.define('product',{
     }
   },
   price: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.INTEGER,
     defaultValue: 0
   },
   productType: {
